requirements.txt
asgiref==3.8.1
Django==5.1.6
django-cors-headers==4.7.0
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
pillow==11.1.0
PyJWT==2.10.1
python-dotenv==1.0.1
sqlparse==0.5.3
tzdata==2025.1
